changes to correct modal delete bug
changing to the modal required using an ng-repeat and an object array for each message.
This is a refactor of that logic to correctly provide an object of confirmationMessages

change from paragraph to ng-repeat

needed to use ng-repeat to list out the items of messages for delete modal

updated dist with recent build.

renamed variable for delete modal ctrl test

added extra condition for variable for test

This makes sure the controller provides a confirmationMessages variable for the view to use

delete & submit buttons focus and enter submit

This is a bit of a hack but ui.bootstrap has issues with respecting html5 autofocus. it does it once and then doesn't work after a modal has been dismissed.
In order to work around this, there's a small timeout implemented when a modal gets brought up to focus it on the main submit element.
form tags were added to get the enter submit working on each modal.

updated dist folder with build

provided better implementation of focus

updated dist folder with build again